# docs/swagger.yaml
components:
  schemas:
...
    Event:
      properties:
        application_end_time:
          format: date-time
          type: string
        application_start_time:
          format: date-time
          type: string
        event_end_time:
          format: date-time
          type: string
        event_start_time:
          format: date-time
          type: string
        game_version:
          $ref: '#/components/schemas/GameVersion'
        id:
          type: integer
        is_current:
          type: boolean
        is_locked:
          type: boolean
        is_main_event:
          type: boolean
        is_public:
          type: boolean
        max_size:
          type: integer
        name:
          type: string
        patch:
          type: string
        teams:
          items:
            $ref: '#/components/schemas/Team'
          type: array
          uniqueItems: false
        uses_medals:
          type: boolean
        waitlist_size:
          type: integer
      required:
      - application_end_time
      - application_start_time
      - event_end_time
      - event_start_time
      - game_version
      - id
      - is_current
      - is_locked
      - is_main_event
      - is_public
      - max_size
      - name
      - teams
      - uses_medals
      - waitlist_size
      type: object
    EventCreate:
      properties:
        application_end_time:
          format: date-time
          type: string
        application_start_time:
          format: date-time
          type: string
        event_end_time:
          format: date-time
          type: string
        event_start_time:
          format: date-time
          type: string
        game_version:
          $ref: '#/components/schemas/GameVersion'
        id:
          type: integer
        is_current:
          type: boolean
        is_locked:
          type: boolean
        is_main_event:
          type: boolean
        is_public:
          type: boolean
        max_size:
          type: integer
        name:
          type: string
        patch:
          type: string
        uses_medals:
          type: boolean
        waitlist_size:
          type: integer
      required:
      - application_end_time
      - application_start_time
      - event_end_time
      - event_start_time
      - game_version
      - max_size
      - name
      - waitlist_size
      type: object
...
    Team:
      properties:
        abbreviation:
          type: string
        allowed_classes:
          items:
            type: string
          type: array
          uniqueItems: false
        color:
          type: string
        discord_role_id:
          type: string
        event_id:
          type: integer
        id:
          type: integer
        name:
          type: string
      required:
      - allowed_classes
      - event_id
      - id
      - name
      type: object
...
    GameVersion:
      enum:
      - poe1
      - poe2
      type: string
      x-enum-varnames:
      - PoE1
      - PoE2
...
    ApplicationStatus:
      enum:
      - applied
      - accepted
      - waitlisted
      - none
      type: string
      x-enum-varnames:
      - ApplicationStatusApplied
      - ApplicationStatusAccepted
      - ApplicationStatusWaitlisted
      - ApplicationStatusNone
...
    EventStatus:
      properties:
        application_status:
          $ref: '#/components/schemas/ApplicationStatus'
        is_team_lead:
          type: boolean
        number_of_signups:
          type: integer
        number_of_signups_before:
          type: integer
        partner_wish:
          type: string
        team_id:
          type: integer
        users_who_want_to_sign_up_with_you:
          items:
            type: string
          type: array
          uniqueItems: false
      required:
      - application_status
      - is_team_lead
      - number_of_signups
      - number_of_signups_before
      type: object
...
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    email: Liberatorist@gmail.com
    name: Liberator
  description: This is the backend API for the BPL project.
  title: BPL Backend API
  version: "2.0"
openapi: 3.1.0
paths:
...
  /events:
    get:
      description: Fetches all events
      operationId: GetEvents
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Event'
                type: array
          description: OK
      security:
      - BearerAuth: []
      tags:
      - event
    put:
      description: Creates or updates an event
      operationId: CreateEvent
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/EventCreate'
                description: Event to create
                summary: event
        description: Event to create
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
          description: Created
      security:
      - BearerAuth: []
      tags:
      - event
  /events/{event_id}:
    delete:
      description: Deletes an event
      operationId: DeleteEvent
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      tags:
      - event
    get:
      description: Fetches an event by id
      operationId: GetEvent
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
          description: OK
      security:
      - BearerAuth: []
      tags:
      - event
...
  /events/{event_id}/duplicate:
    post:
      description: Duplicates an event's configuration
      operationId: DuplicateEvent
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/EventCreate'
                description: Event to create
                summary: event
        description: Event to create
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
          description: Created
      security:
      - BearerAuth: []
      tags:
      - event
...
  /events/{event_id}/status:
    get:
      description: Gets the status for an event including the user's application status
      operationId: GetEventStatus
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventStatus'
          description: OK
      security:
      - BearerAuth: []
      tags:
      - event
...